+++ http/src/request/guild/create_guild/mod.rs
@@ -191,7 +191,7 @@ pub struct TextFields {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub parent_id: Option<Id<ChannelMarker>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub rate_limit_per_user: Option<u64>,
+    pub rate_limit_per_user: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub topic: Option<String>,
 }
